+++ controller/konnect/reconciler_generic.go
@@ -484,6 +484,17 @@ func (r *KonnectEntityReconciler[T, TEnt]) Reconcile(ctx context.Context, ent TE
 		return patchWithProgrammedStatusConditionBasedOnOtherConditions(ctx, r.Client, ent)
 	}
 
+	// If a type has a KonnectConfigStore ref (KongVault), handle it.
+	var resolvedConfigStoreID string
+	res, stop, resolvedConfigStoreID, err = handleConfigStoreRef(ctx, r.Client, ent)
+	if err != nil || !res.IsZero() {
+		return res, err
+	}
+	if stop {
+		return patchWithProgrammedStatusConditionBasedOnOtherConditions(ctx, r.Client, ent)
+	}
+	ctx = ops.WithResolvedConfigStoreID(ctx, resolvedConfigStoreID)
+
 	programmedFalseCondition := metav1.Condition{
 		Type:    konnectv1alpha1.KonnectEntityProgrammedConditionType,
 		Status:  metav1.ConditionFalse,